Release 1.6.3 (2008-12-21)
- Fixed SEG fault for 32-bit Linux clients on top of 64-bit kernels
  Thanks to Jani Ollikainen (pronetko.fi) for reporting and debugging this.
  Thanks to Thomas Rzipa for reporting this issue, too.
- Fixed setqlim to allow setting limits exceeding 2^32 blocks or files.
  The current solution will work only for perl executables which support
  64-bit wide integers (e.g. x86-64)

+++ Quota/hints/linux.h
@@ -40,7 +40,7 @@
 #define QCMD(cmd, type)  (((cmd) << SUBCMDSHIFT) | ((type) & SUBCMDMASK))
 
 /* declare an internal version of the quota block struct */
-typedef unsigned int qsize_t;
+typedef u_int64_t qsize_t;
 struct dqblk {
   qsize_t dqb_ihardlimit;   /* absolute limit on allocated inodes */
   qsize_t dqb_isoftlimit;   /* preferred inode limit */
